Xcode: Replace PowerPlant files with open-powerplant files
When Xcode imported the CodeWarrior project, all of the PowerPlant files
were referenced from the system CodeWarrior tree (this is added as a "Source Tree"
and is a global Xcode preference).

Use the open-powerplant files instead, and remove all the files referenced from CodeWarrior.
Also add open-powerplant as a (recursive) include directory, and add PP_Xcode.pch as the
precompiled prefix header.

This project should actually compile (although with some errors) on Xcode 2.5/OS X 10.4!