Xcode: Remove extra boost headers from xcodeproj
When importing the project, Xcode went through the dependencies of each file and
added them to the project in the "Imported Files" group.

Tthe files left in that group are (lots of) boost headers which are all referenced
by absolute file paths (which would need to be fixed). I can't see any reason to leave
these headers in the project at all, they're found fine through the search path.

Remove them and the "Imported Files" group.