Xcode: Update project file for Xcode 3.2 (OS X 10.6)
Xcode 3.2 is the earliest Xcode version available for OS X 10.6, open
the Xcode project with 3.2 and let it make its changes.

Only intentional change is fixing the GCC version to 4.0 (it's required
to compile against the 10.4 SDK).

Note that binaries built by Xcode 3.2 are 4.8 MB vs 5.9 MB from Xcode 3.1
on 10.5. I don't know why, but the binary seems to work correctly.